Move GenerateKey function out of the worker file
